UI and logic corrections

+++ lib/Api/BusinessPage/attach_new_invoice.dart
@@ -760,8 +760,8 @@ class _AttachNewInvoiceState extends State<AttachNewInvoice> {
         });
   }
 
-  int calculateTotal() {
-    int total = TotalAndSubTotal().getSubTotal(context,isNewInvoice);
+  double calculateTotal() {
+    double total = TotalAndSubTotal().getSubTotal(context,isNewInvoice);
 
     if(_children.length == 0){
