+++ lib/Api/BusinessPage/update_invoice.dart
@@ -764,8 +764,8 @@ class _UpdateInvoiceDataState extends State<UpdateInvoiceData> {
               });
         });
   }
-  int calculateTotal() {
-    int total = TotalAndSubTotal().getSubTotal(context, isNewInvoice);
+  double calculateTotal() {
+    double total = TotalAndSubTotal().getSubTotal(context, isNewInvoice);
 
     if(_children.length == 0){
